third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-override-self.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    width: 100px;
    height: 100px;

    --foo: {
        --foo: green;
    };

    @apply --foo;
    background-color: var(--foo);
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).getPropertyValue("--foo"), " green");
}, "@apply can override the custom property it applies");
</script>
