third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-override-self-expected.txt

This is a testharness.js-based test.
FAIL @apply can override the custom property it applies assert_equals: expected "rgb(0, 128, 0)" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
